% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityfuncs.R
\name{BLP_eventtimes}
\alias{BLP_eventtimes}
\title{BLP Event Times Analysis}
\usage{
BLP_eventtimes(data = data, nperiods = 2, rhs_formula = NULL)
}
\arguments{
\item{data}{The dataset to be used in the analysis, output from the function het_prep}

\item{nperiods}{The number of periods to consider in the analysis. Note this must be fewer than the number
of periods where e >=0}

\item{rhs_formula}{A string representing the RHS of the linear model formula, including
desired covariates. Default is NULL.}
}
\value{
A list of `nperiods` elements, each containing the summary of the linear model
fitted for a specific event time.
}
\description{
This function facilitates BLP analysis with a flexible
RHS formula, allowing users to input any desired covariates. The function iteratively
fits a linear model for different event times (`e.times`) in the data.
}
\details{
The formula for the linear model includes an intercept (denoted by `1` in the formula). Users must
provide the RHS formula as a string, which will be appended to `dynamic.cate ~ 1 +` to
form the full model formula.
}
\examples{
# Assuming data is predefined and has appropriate structure:
results_list <- BLP_eventtimes(data, nperiods = 3, rhs_formula = "X1 + X2 + X3")

}
