% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatesimdata.R
\name{build_sim_dataset}
\alias{build_sim_dataset}
\title{Build Simulation Dataset}
\usage{
build_sim_dataset(
  time.periods = 4,
  n = 2500,
  random = TRUE,
  noise = FALSE,
  time.dependent.covar = TRUE,
  confounding = 1,
  het = "none",
  chi = 1,
  taumodel = 1,
  verbose = FALSE
)
}
\arguments{
\item{time.periods}{An integer specifying the number of time periods in the dataset.}

\item{n}{An integer indicating the number of observations to generate.}

\item{random}{Logical flag to indicate random assignment (TRUE) or confounding (FALSE) in treatment assignment.}

\item{noise}{NOT IMPLEMENTED YET Logical flag to add (TRUE) or not add (FALSE) 100 extra noise covariates.TO BE FIXED}

\item{time.dependent.covar}{Logical flag to include (TRUE) or exclude (FALSE) a time-dependent covariate.}

\item{confounding}{An integer specifying the complexity of confounding in the model. 
1 indicates simple confounding depending on one covariate, alternative values indicate more complex confounding structures.}

\item{het}{Character string indicating the type of heterogeneity in the treatment effect. 
Possible values are "none" (no heterogeneity), "constant", or "calendar".}

\item{chi}{An integer specifying how many covariates have a beta effect.}

\item{taumodel}{An integer to choose the model for treatment effect heterogeneity.}

\item{verbose}{whether to print values of various params, default is FALSE}
}
\value{
A generated dataset as an R dataframe, structured for use in MLDID.G indicates group membership, y0
and y1 are the true potential outcomes, "delta.e" is time since event, and "G1"... indicate probability 
of being in each group.
}
\description{
This function generates a simulated dataset. It creates a panel data structure 
with both treated and untreated potential outcomes over specified time periods. 
The function allows customization of confounding, heterogeneity, and noise factors in the data. 
There are no observations which are treated in the first time period.
}
\details{
Code for simulating data, adapted from CSA

Julia Hatamyar Began 31 August 2022

Additions Feb/March 2023: time varying X5
Additions APril 2023: noise covariates 
Additions Jan 2024: package formatting, clarification
==========================================
}
