% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityfuncs.R
\name{CLAN_glhtest}
\alias{CLAN_glhtest}
\title{Generalized Linear Hypothesis Test in CLAN Analysis}
\usage{
CLAN_glhtest(data = data, affected = affected, thres = 0.2, alpha = 0.05)
}
\arguments{
\item{data}{The dataset containing variables for analysis, output of function het_prep}

\item{affected}{A vector of variables for which the effects are to be tested.}

\item{thres}{A threshold value used to determine the high and low effect based on
dynamic.cate quantiles.}

\item{alpha}{Significance level used for confidence intervals.}
}
\value{
A three-dimensional array with dimensions corresponding to the number of
affected variables, the number of statistical measures (coefficients, CIs, one-sided p-values),
and the number of event times.
}
\description{
Adapted from the methodology in Demirer and Chernozhukov (2018), this function
performs a generalized linear hypothesis (GLH) test for each event time in the CLAN
(Classification Analysis) framework. It computes effects, confidence intervals,
and one-sided test p-values for specified variables based on a quantile-based
threshold, using quantiles of the estimated CATT.
}
\examples{
# Assuming data and affected are predefined:
results_array <- CLAN_glhtest(data, affected, thres = 0.2, alpha = 0.05)

}
