% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityfuncs.R
\name{CLAN_ttest}
\alias{CLAN_ttest}
\title{T-Test Analysis for CLAN}
\usage{
CLAN_ttest(data = data, affected = affected)
}
\arguments{
\item{data}{The dataset containing variables for analysis.}

\item{affected}{A vector of variables for which the t-tests are to be conducted.}
}
\value{
A three-dimensional array with dimensions corresponding to the number of
affected variables, the number of statistical measures (means, CIs, p-values),
and the number of event times.
}
\description{
This function, part of the CLAN framework, performs
t-tests to compare the means of affected variables between most and least effected groups.
These groups are determined based on dynamic.cate quantiles. The function operates
for each event time separately.
}
