% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterloop.R
\name{didMLloop_oracle}
\alias{didMLloop_oracle}
\title{Obtaining oracle effects for simulations}
\usage{
didMLloop_oracle(dta = dta, xformla = xformla)
}
\arguments{
\item{dta}{A data frame containing the simulation data. This should include variables for
treatment status, subgroup membership, and potential outcomes under treatment and control.}

\item{xformla}{A formula or other specification of the model used in the simulation,
indicating how the covariates are related to the outcomes.}
}
\value{
An object of class "MLDID" containing the estimated treatment effects and other
        relevant information. This object includes the following components:
        - attgt: A list of estimated treatment effects for each group and time period.
        - cates: A matrix of Conditional Average Treatment Effects on the Treated.
        - positions: A matrix indicating the positions of each unit in the data.
        - IDs: A matrix of IDs for each unit in the data.
        - params: A list of parameters used in the simulation, including data, time lists,
                 group lists, and number of periods, groups, and units.
}
\description{
This obtains the true ATTs and CATTs by direct calculation, using the known values
of the potential outcomes under treatment and control.
}
\examples{
# Example usage:
# didMLloop_oracle(dta = your_data, xformla = your_formula)

}
