% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{dynamic_attgt}
\alias{dynamic_attgt}
\title{Aggregate ATTs for Dynamic Event Study}
\usage{
dynamic_attgt(attgt.result)
}
\arguments{
\item{attgt.result}{An object of class "MLDID" containing the ATT results to be aggregated.}
}
\value{
A list containing the processed "MLDID" object, dynamic ATT estimates, standard errors, 
        covariate-specific ATTs, and other relevant data used in the computation.
}
\description{
This function aggregates average treatment effects on the treated (ATTs) to create a dynamic 
"event study" type analysis. It takes an "MLDID" object as input, processes the results, and 
computes ATTs specific to each event time.
}
\details{
Aggregating ATTs and CATES/scores 
Julia Hatamyar, new file 11 May 2023, original code earlier version  

========================================
}
