% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{dynamic_attgt_or}
\alias{dynamic_attgt_or}
\title{Aggregate Oracle ATTs}
\usage{
dynamic_attgt_or(data = data, attgt.res = attgt.res)
}
\arguments{
\item{data}{A data frame}

\item{attgt.res}{An object containing oracle ATT results 
with group-time information.}
}
\value{
A list containing the aggregated oracle ATT estimates and other relevant data 
        used in the computation.
}
\description{
This function aggregates oracle average treatment effects on the treated (ATTs) based on 
the provided data. It is designed for processing and aggregating ATTs in a dynamic event study 
context, using oracle or known ATTs.
}
