% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{dynamic_cates}
\alias{dynamic_cates}
\title{Structure CATES Dynamically}
\usage{
dynamic_cates(res, type = c("cates", "scores"), oracle = FALSE)
}
\arguments{
\item{res}{An object of class "MLDID" containing the results from which CATES will be structured dynamically.}

\item{type}{A character vector specifying the type of data to be structured. 
Options are "cates" for Conditional Average Treatment Effects or "scores" for double-robust scores.
Default is "cates".}

\item{oracle}{A logical flag indicating whether the function is being used for simulations 
(oracle = TRUE) or real data analysis (oracle = FALSE). Default is FALSE.}
}
\value{
A data frame containing the dynamically structured CATES or scores, with each row 
        corresponding to a unique identifier from the input data and columns for each event time.
}
\description{
This function structures Conditional Average Treatment Effects (CATES) dynamically based on 
time to events. It takes an object of class "MLDID" as input and processes it to compute dynamic CATES.
}
