% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityfuncs.R
\name{het_prep}
\alias{het_prep}
\title{Prepare Heterogeneous Treatment Effects for Analysis}
\usage{
het_prep(attgt.result, cates)
}
\arguments{
\item{attgt.result}{The output of the MLDID function.}

\item{cates}{The output of the dynamic_cates function, which could be CATEs or scores.}
}
\value{
A panel data frame where each row represents a unit (e.g., an individual or a firm)
at a specific event time, merged with the corresponding CATE or score.
}
\description{
Merges dynamic conditional average treatment effects (CATEs) or scores
with the original dataset by event-time. It is  used for conducting BLP
or CLAN analyses.
}
