% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATTestimator.R
\name{LNW_DiD}
\alias{LNW_DiD}
\title{Compute the Treatment Effects}
\usage{
LNW_DiD(
  X,
  Y,
  Ti,
  Si,
  constant_eff = c("constant", "non_constant"),
  gamma = NULL,
  k_folds = 10,
  lambda_choice = "lambda.min",
  penalty_factor = NULL,
  tune_penalty = TRUE,
  nu_model = "rlearner",
  sigma_model = "rlearner",
  delta_model = "glm",
  t_func = FALSE
)
}
\arguments{
\item{X}{A data frame of covariates.}

\item{Y}{A numeric vector of outcomes.}

\item{Ti}{A numeric vector indicating time.}

\item{Si}{A numeric vector indicating subgroup membership.}

\item{constant_eff}{A character vector indicating whether the effect is constant or non-constant. 
Accepted values are "constant" or "non_constant".}

\item{gamma}{Optional numeric value specifying the gamma parameter. If NULL, the default TMLE is used.}

\item{k_folds}{An integer indicating the number of folds for cross-validation. Default is 10.}

\item{lambda_choice}{A character string indicating the method for choosing lambda in the LASSO model. 
Accepted values are "lambda.min" or "lambda.1se". Default is lambda.min}

\item{penalty_factor}{Optional numeric vector for penalty factors in the LASSO model.}

\item{tune_penalty}{A logical indicating whether to tune the penalty factors. Default is TRUE.}

\item{nu_model}{A character string specifying the model to use for nu estimation. 
Accepted values are "rlearner" and "cf" (for causal forest).}

\item{sigma_model}{A character string specifying the model to use for sigma estimation. 
Accepted values are "rlearner" and "cf" (for causal forest).}

\item{delta_model}{A character string specifying the model to use for delta estimation. 
Accepted values are "glm" and "SL" (for SuperLearner).}

\item{t_func}{A logical indicating whether to use a functional form for t. Default is FALSE.}
}
\value{
A list containing the estimated treatment effects and associated statistics, 
        including TAU_hat, standard error estimates, tau_hat, score estimates, and y estimates.
}
\description{
This function runs machine learning for group-time ATTs (Average Treatment Effects on the Treated) 
and CATTs (Conditional Average Treatment Effects on the Treated) as implemented in the methodology 
of Lu, Nie, and Wager. It allows for various configurations and ML models to estimate treatment effects.
}
\details{
==========================================
}
\examples{
# Example usage:
# LNW_DiD(X = your_covariates, Y = your_outcomes, Ti = your_treatment_indicator, 
#         Si = your_subgroup_indicator)

}
