% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLDID_class.R, R/masterloop.R
\docType{class}
\name{MLDID}
\alias{MLDID}
\title{MLDID Class}
\usage{
MLDID(
  outcome,
  group,
  time,
  id_name,
  data,
  xformla = xformla,
  tune_penalty = TRUE,
  nu_model = "rlearner",
  sigma_model = "rlearner",
  delta_model = "glm",
  t_func = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{outcome}{Name of the outcome variable in the data.}

\item{group}{Name of the group variable in the data, typically indicating treatment assignment.}

\item{time}{Name of the time variable in the data.}

\item{id_name}{Name of the identifier variable in the data.}

\item{data}{The dataframe to be used for the analysis.}

\item{xformla}{A formula object specifying the covariates to be used in the model.}

\item{tune_penalty}{Logical flag indicating whether to tune the penalty parameters (TRUE) or not (FALSE).
Parameters to be tuned are in nu, sigma, and delta - see the function LNW_DID for details}

\item{nu_model}{Model to be used for the nu component. Possible values are "rlearner" (default) or "cf" for causal forest.}

\item{sigma_model}{Model to be used for the sigma component. Possible values are "rlearner" (default) or "cf" for causal forest.}

\item{delta_model}{Model to be used for the delta component. Possible values are "glm" (default) or "SL" for superlearner.}

\item{t_func}{A function specifying whether to estimate t_hat. Default is (TRUE), if (FALSE), will replace with 0.5, since there is always one observation pre/post
NOTE that you MUST specify this as FALSE if there are no time-varying covariates in the data}

\item{verbose}{Logical flag indicating whether to print detailed information during function execution (default is TRUE).}
}
\value{
An object of class `MLDID` containing the estimated treatment effects and other
        relevant information. This object includes the following components:
        - attgt: A list of estimated treatment effects for each group and time period, using both the MLDID
        and DRDID (Callaway & Sant'Anna) estimators. The MLDID estimates are 'TAU-hat' and the DRDID are 'attgt'.
        - cates: A matrix of Conditional Average Treatment Effects on the Treated.
        - scores: A matrix of double-robust scores for the Treated.
        - gammas: A matrix of estimated gammas
        - positions: A matrix indicating the positions of each unit in the data.
        - IDs: A matrix of IDs for each unit in the data.
        - params: A list of parameters used in the simulation, including data, time lists,
                 group lists, and number of periods, groups, and units.
}
\description{
This class is for storing the output of the main MLDID estimating procedure,
and is passed to follow-on functions. .

runs the machine learning for the group-time ATTs and CATTs. See README for example usage.
}
\details{
==========================================
}
\section{Slots}{

\describe{
\item{\code{attgt}}{group-time ATTs}

\item{\code{cates}}{Conditional Average Treatment Efffects}

\item{\code{scores}}{double-robust scores}

\item{\code{gammas}}{gammas used in the weighting procedure (Lu Nie Wager)}

\item{\code{positions}}{positions of the variable (TO REMOVE)}

\item{\code{IDs}}{Identifier of the observations}

\item{\code{params}}{Parameters used by following functions}

\item{\code{processed_attgt}}{Space for post-processed version of the estimated results}
}}

