% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityfuncs.R
\name{plotBLP}
\alias{plotBLP}
\title{Plot BLP Coefficients Over Event Time}
\usage{
plotBLP(BLP_output, affected)
}
\arguments{
\item{BLP_output}{A list containing the output of `BLP_eventtimes`.}

\item{affected}{A character vector of the names of the variables included in the BLP model
analysis (excluding the intercept).}
}
\value{
A ggplot object representing the coefficient estimates and their confidence intervals
for each variable across different event times. Points are slightly dodged for readability,
and the plot displays integer values for event times on the x-axis.
}
\description{
Creates a plot of coefficient estimates and their confidence intervals
for various variables over different event times. The plot is generated based on
the output from the BLP_eventtimes function. It excludes the
intercept and presents each variable with a slight horizontal offset for readability.
}
\examples{
# Assuming BLP_output and affected are predefined:
plot <- plotBLP(BLP_output, affected)
plot

}
