% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuncs.R
\name{process_attgt}
\alias{process_attgt}
\title{Process ATTGT Results}
\usage{
process_attgt(attgt.result)
}
\arguments{
\item{attgt.result}{An object of class "MLDID" containing the results to be processed.}
}
\value{
An updated "MLDID" object that includes the processed ATT results and other
        relevant statistics, such as standard errors and covariate-specific ATTs.
}
\description{
This function processes the results stored in an "MLDID" object, extracting and organizing
Average Treatment Effects on the Treated (ATT) estimates and related statistics.Usually this
will be performed as an intermediate step within the aggregation function but it might be
useful to see the group-time ATTs on their own.
}
\examples{
# Example usage:
# processed_result <- process_attgt(your_MLDID_object)

}
