% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuncs.R
\name{process_attgt_sim}
\alias{process_attgt_sim}
\title{Process ATTGT Simulation Results}
\usage{
process_attgt_sim(attgt.result)
}
\arguments{
\item{attgt.result}{An updated object of class "MLDID" containing the simulation results to be processed.}
}
\value{
An updated "MLDID" object that includes the processed simulation-specific ATT results,
        along with the grouping and time period information.
}
\description{
This function is specifically tailored for processing simulation results stored in an "MLDID"
object, which has already been run through the non-simulated version. It focuses on extracting and organizing simulation-specific ATT estimates.
}
\examples{
# Example usage:
# processed_sim_result <- process_attgt_sim(your_MLDID_object)

}
